requirements.txt
discord.py>=2.5.2
aiosqlite>=0.19.0
python-dotenv>=1.0.0
captcha>=0.7.1
Pillow>=10.0.0
feedparser>=6.0.0
aiohttp>=3.8.0 